requirements.txt
asgiref==3.3.4
Django==3.2.3
gunicorn==20.1.0
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1

djangorestframework~=3.12.4
cffi~=1.14.5
Pillow~=8.2.0
pip~=21.0.1
google~=3.0.0
Jinja2~=2.11.3
setuptools~=49.2.1
uritemplate~=3.0.1
google-api-python-client~=2.4.0
requests~=2.25.1